+++ CHANGES.md
@@ -229,6 +229,8 @@ Tools
 
 - Coq options can be set on the command line, eg `-set "Universe Polymorphism=true"`
 
+- coq_makefile's install target now errors if any file to install is missing.
+
 Standard Library
 
 - Added lemmas about monotonicity of `N.double` and `N.succ_double`, and about